Oligomer Display: Decided on Pearson correlation over Spearman correlation because of speed and lack of improvement.
git-svn-id: https://skittle.svn.sourceforge.net/svnroot/skittle@108 6d041a4c-438f-4d6f-bb7b-eaab962e81b8

+++ branches/newui/BasicTypes.h
@@ -69,11 +69,11 @@ public:
 
 class point{
 public:
-	float x;
-	float y;
-	float z;
+	double x;
+	double y;
+	double z;
 	point(){x=0; y=0; z=0;}
-	point(float X, float Y, float Z){
+	point(double X, double Y, double Z){
 		x=X;
 		y=Y;
 		z=Z;
